Add basic support for DW_OP_entry_value, DW_OP_const_type, DW_OP_implicit_value
These ops are now parsed and encoded correctly, but `entry_value` definitely
need more support from the assembler. It essentially needs to embed another
expression as its operand. So it would be more useful to a user to change from:

    DW_OP_entry_value(<size-of-expression>, <encoded expression>)

to

    DW_OP_entry_value(<expression>)

Perhaps we want a non-dwarf operation for that. E.g.

    SINISTER_entry_value(<expression>)